+++ packages/antd-ext/components/ButtonExt/ButtonExt.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, useContext } from "react";
+import React, { useCallback, useState } from "react";
 import { Button } from "antd";
 import type { ButtonProps } from "antd";