+++ packages/antd-ext/components/ButtonExt/index.ts
@@ -1,3 +1,7 @@
 import Mix from './Mix';
+import BaseButtonExt from './ButtonExt';
 
 export default Mix;
+export {
+  BaseButtonExt
+}