+++ packages/antd-ext/components/DrawerExt/index.ts
@@ -1,3 +1,7 @@
 import Mix from './Mix';
+import BaseDrawerExt from './DrawerExt';
 
 export default Mix;
+export {
+  BaseDrawerExt,
+}