+++ packages/antd-ext/components/InputExt/index.ts
@@ -1,4 +1,5 @@
 import { MixinInputExt, MixinInputOutLineExt } from './Mix';
 import InputOutLineExt from './InputOutLineExt';
+import { InputExt as BaseInputExt } from './InputExt';
 
-export { InputOutLineExt, MixinInputExt, MixinInputOutLineExt };
+export { InputOutLineExt, MixinInputExt, MixinInputOutLineExt, BaseInputExt };