+++ packages/antd-ext/components/LayoutExt/index.ts
@@ -1,3 +1,7 @@
 import Mixin from "./Mix";
+import BaseLayoutExt from './LayoutExt';
 
 export default Mixin;
+export {
+  BaseLayoutExt,
+}