+++ packages/antd-ext/components/RangeNumberExt/index.ts
@@ -1,5 +1,6 @@
 import RangeNumberExt from './RangeNumberExt';
 import RangeNumberOutLineExt from './RangeNumberOutLineExt';
 import { MixinRangeNumberExt, MixinRangeNumberOutLineExt } from './Mix';
+import BaseRangeNumberExt from './RangeNumberExt';
 
-export { RangeNumberExt, RangeNumberOutLineExt, MixinRangeNumberExt, MixinRangeNumberOutLineExt };
+export { RangeNumberExt, RangeNumberOutLineExt, MixinRangeNumberExt, MixinRangeNumberOutLineExt, BaseRangeNumberExt };