+++ packages/antd-ext/components/TableExt/index.ts
@@ -1,3 +1,7 @@
 import Mix from './Mix';
+import BaseTableExt from './TableExt';
 
 export default Mix;
+export {
+  BaseTableExt
+}