+++ packages/antd-ext/components/index.ts
@@ -1,17 +1,26 @@
-import ButtonExt from './ButtonExt';
+import ButtonExt, { BaseButtonExt } from './ButtonExt';
 import ConfigProviderExt from './ConfigProviderExt';
-import TableExt from './TableExt';
-import { MixinInputExt as InputExt, MixinInputOutLineExt as InputOutLineExt } from './InputExt';
-import { MixinSelectExt as SelectExt, MixinSelectOutLineExt as SelectOutLineExt } from './SelectExt';
-import DrawerExt from './DrawerExt';
-import { MixinRangeNumberExt as RangeNumberExt, MixinRangeNumberOutLineExt as RangeNumberOutLineExt } from './RangeNumberExt';
-import LayoutExt from './LayoutExt';
+import TableExt, { BaseTableExt } from './TableExt';
+import { MixinInputExt as InputExt, MixinInputOutLineExt as InputOutLineExt, BaseInputExt } from './InputExt';
+import { MixinSelectExt as SelectExt, MixinSelectOutLineExt as SelectOutLineExt, BaseSelectExt } from './SelectExt';
+import DrawerExt, { BaseDrawerExt } from './DrawerExt';
+import { MixinRangeNumberExt as RangeNumberExt, MixinRangeNumberOutLineExt as RangeNumberOutLineExt, BaseRangeNumberExt } from './RangeNumberExt';
+import LayoutExt, { BaseLayoutExt } from './LayoutExt';
+import DatePickerExt from './DatePickerExt';
 export * from 'antd';
-import zh_CN from 'antd/locale/zh_CN';
-import en_US from 'antd/locale/en_US';
+import zh_CN from 'antd/es/locale/zh_CN';
+import en_US from 'antd/es/locale/en_US';
 import type { Theme } from './ConfigProviderExt/context';
 
 export {
+  BaseButtonExt,
+  BaseTableExt,
+  BaseInputExt,
+  BaseSelectExt,
+  BaseDrawerExt,
+  BaseRangeNumberExt,
+  BaseLayoutExt,
+
   ButtonExt,
   ConfigProviderExt,
   TableExt,
@@ -23,6 +32,7 @@ export {
   RangeNumberExt,
   RangeNumberOutLineExt,
   LayoutExt,
+  DatePickerExt,
 
   Theme,
 }